+++ client/src/types/data/IMyCollection.ts
@@ -1,6 +1,8 @@
 import IEntity from './IEntity'
+import IEvent from './IEvent'
 import ISet from './ISet'
 
 export default interface IMyCollection extends ISet {
   containing?: Array<IEntity>
+  added_to_by?: Array<IEvent>
 }